New codebase
Complete revamp of codebase. Polyglot is now being developed on Windows
WAMP stack using phpmyadmin. There is now an encoding issue with
phpmyadmin and Japanese characters, which has been partially resolved by
using the Big5 character set. Also, the slideshow on the login page is a
bit choppy in places. Not sure as to why this is.